"refactor/채팅관련 타입 Types.ts로 이동"
